All work sectors separated to specific namespaces. IniFile folder renames to SetupFiles directory and moved to local files.
