--- TCPSharpFileSync/IniFile/setup1.ini
@@ -1,14 +1,0 @@
-﻿[General]
-directoryPath = dir
-port = 9000
-msTimeout = 70000000
-hashDictionaryName = name
-
-[Client]
-ip = ip.ip.ip.ip
-doDownload = Yes
-doUpload = Yes
-removeIfNotOnClient = No
-removeIfNotOnServer = Yes
-
-[Server]
